requirements.txt
asyncio
requests
websockets
python-telegram-bot==20.0
